Dashboard

+++ src/main/java/com/miniproject/eventure/infrastructure/user/dto/CreateUserPointsRequestDTO.java
@@ -5,6 +5,7 @@ import com.miniproject.eventure.entity.user.UserPoints;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 
@@ -13,7 +14,7 @@ import java.time.OffsetDateTime;
 public class CreateUserPointsRequestDTO {
 
     @NotNull
-    private Double points;
+    private BigDecimal points;
 
     private String validityPeriod;
 
@@ -22,7 +23,7 @@ public class CreateUserPointsRequestDTO {
 
         userPoints.setUser(user);
         userPoints.setPoints(this.points);
-        userPoints.setUsedPoints(0.0);
+        userPoints.setUsedPoints(BigDecimal.valueOf(0.0));
         userPoints.setExpiredAt(expiryAt);
 
         return userPoints;
